Final Version Completed

+++ rolt/src/main/java/dev/mikita/rolt/rest/PropertyController.java
@@ -7,6 +7,7 @@ import dev.mikita.rolt.entity.*;
 import dev.mikita.rolt.exception.NotFoundException;
 import dev.mikita.rolt.exception.ValidationException;
 import dev.mikita.rolt.rest.util.RestUtils;
+import dev.mikita.rolt.security.model.CustomUserDetails;
 import dev.mikita.rolt.service.CityService;
 import dev.mikita.rolt.service.LandlordService;
 import dev.mikita.rolt.service.PropertyService;
@@ -21,14 +22,19 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
+import java.security.Principal;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+/**
+ * The type Property controller.
+ */
 @RestController
 @RequestMapping("/rest/v1/properties")
 public class PropertyController {
@@ -38,6 +44,13 @@ public class PropertyController {
     private final LandlordService landlordService;
     private final CityService cityService;
 
+    /**
+     * Instantiates a new Property controller.
+     *
+     * @param propertyService the property service
+     * @param landlordService the landlord service
+     * @param cityService     the city service
+     */
     @Autowired
     public PropertyController(PropertyService propertyService,
                               LandlordService landlordService,
@@ -47,14 +60,26 @@ public class PropertyController {
         this.cityService = cityService;
     }
 
+    /**
+     * Gets properties.
+     *
+     * @param page         the page
+     * @param size         the size
+     * @param cityId       the city id
+     * @param propertyType the property type
+     * @param minSquare    the min square
+     * @param maxSquare    the max square
+     * @param isAvailable  the is available
+     * @return the properties
+     */
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Map<String, Object>> getProperties(
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "20") int size,
             @RequestParam(required = false) Integer cityId,
             @RequestParam(required = false) PropertyType propertyType,
-            @RequestParam(required = false) Integer minSquare,
-            @RequestParam(required = false) Integer maxSquare,
+            @RequestParam(required = false) Double minSquare,
+            @RequestParam(required = false) Double maxSquare,
             @RequestParam(required = false) Boolean isAvailable) {
 
         ModelMapper modelMapper = new ModelMapper();
@@ -84,6 +109,12 @@ public class PropertyController {
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
+    /**
+     * Gets property.
+     *
+     * @param id the id
+     * @return the property
+     */
     @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponsePublicPropertyDto getProperty(@PathVariable Integer id) {
         final Property property = propertyService.find(id);
@@ -93,9 +124,24 @@ public class PropertyController {
         return new ModelMapper().map(property, ResponsePublicPropertyDto.class);
     }
 
-//    @PreAuthorize("hasRole('ROLE_LANDLORD')")
+    /**
+     * Create property response entity.
+     *
+     * @param principal the principal
+     * @param propertyDto the property dto
+     * @return the response entity
+     */
+    @PreAuthorize("hasAnyRole('ROLE_LANDLORD', 'ROLE_ADMIN', 'ROLE_MODERATOR')")
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Void> createProperty(@RequestBody @Valid RequestCreatePropertyDto propertyDto) {
+    public ResponseEntity<Void> createProperty(Principal principal, @RequestBody @Valid RequestCreatePropertyDto propertyDto) {
+        final CustomUserDetails userDetails = (CustomUserDetails) principal;
+        final User user = userDetails.getUser();
+
+        if ((user.getRole() != Role.ADMIN || user.getRole() != Role.MODERATOR)
+                && !user.getId().equals(propertyDto.getOwnerId())) {
+            throw new AccessDeniedException("You cannot create properties for other users.");
+        }
+
         ModelMapper modelMapper = new ModelMapper();
         modelMapper.getConfiguration().setAmbiguityIgnored(true);
 
@@ -117,10 +163,21 @@ public class PropertyController {
         return new ResponseEntity<>(headers, HttpStatus.CREATED);
     }
 
-//    @PreAuthorize("hasAnyRole('ROLE_LANDLORD', 'ROLE_LANDLORD', 'ROLE_MODERATOR')")
+    /**
+     * Update property.
+     *
+     * @param principal the principal
+     * @param id          the id
+     * @param propertyDto the property dto
+     */
+    @PreAuthorize("hasAnyRole('ROLE_LANDLORD', 'ROLE_LANDLORD', 'ROLE_MODERATOR')")
     @PutMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void updateProperty(@PathVariable Integer id, @RequestBody @Valid RequestUpdatePropertyDto propertyDto) {
+    public void updateProperty(
+            Principal principal,
+            @PathVariable Integer id,
+            @RequestBody @Valid RequestUpdatePropertyDto propertyDto) {
+
         final Property original = propertyService.find(id);
         if (original == null) {
             throw NotFoundException.create("Property", id);
@@ -128,6 +185,14 @@ public class PropertyController {
 
         if (!original.getId().equals(propertyDto.getId())) {
             throw new ValidationException("Property identifier in the data does not match the one in the request URL.");
+        }
+
+        final CustomUserDetails userDetails = (CustomUserDetails) principal;
+        final User user = userDetails.getUser();
+
+        if ((user.getRole() != Role.ADMIN || user.getRole() != Role.MODERATOR)
+                && !user.getId().equals(original.getOwner().getId())) {
+            throw new AccessDeniedException("You cannot update properties for other users.");
         }
 
         final City city = cityService.find(propertyDto.getCityId());
@@ -144,13 +209,29 @@ public class PropertyController {
         LOG.debug("Updated property {}.", propertyDto);
     }
 
+    /**
+     * Delete property.
+     *
+     * @param principal the principal
+     * @param id the id
+     */
     @DeleteMapping(value = "/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteProperty(@PathVariable Integer id) {
+    @PreAuthorize("hasAnyRole('ROLE_LANDLORD', 'ROLE_LANDLORD', 'ROLE_MODERATOR')")
+    public void deleteProperty(Principal principal, @PathVariable Integer id) {
         final Property toRemove = propertyService.find(id);
         if (toRemove == null) {
             return;
         }
+
+        final CustomUserDetails userDetails = (CustomUserDetails) principal;
+        final User user = userDetails.getUser();
+
+        if ((user.getRole() != Role.ADMIN || user.getRole() != Role.MODERATOR)
+                && !user.getId().equals(toRemove.getOwner().getId())) {
+            throw new AccessDeniedException("You cannot delete properties for other users.");
+        }
+
         propertyService.remove(toRemove);
         LOG.debug("Removed property {}.", toRemove);
     }
